move swap_ambiguous_atom_thin_atoms into free function
