+++ src/beignet/structure/_residue_array.py
@@ -2,7 +2,6 @@ import dataclasses
 import functools
 import io
 import operator
-from functools import cache
 from typing import Callable, Literal
 
 import einops
@@ -17,7 +16,7 @@ from optree.dataclasses import dataclass
 from torch import Tensor
 
 from beignet import pad_to_target_length
-from beignet.constants import AMINO_ACID_1_TO_3, ATOM_THIN_ATOMS, STANDARD_RESIDUES
+from beignet.constants import ATOM_THIN_ATOMS, STANDARD_RESIDUES
 
 from ._atom_array_to_atom_thin import atom_array_to_atom_thin
 from ._atom_thin_to_atom_array import atom_thin_to_atom_array
@@ -45,29 +44,6 @@ def _atom_name_mask(atom_name: str, device=None) -> Tensor:
             for v in ATOM_THIN_ATOMS.values()
         ]
     )
-
-
-# 180 degree symmetry
-_AMBIGUOUS_ATOM_SWAPS = {
-    "ASP": {"OD1": "OD2"},
-    "GLU": {"OE1": "OE2"},
-    "PHE": {"CD1": "CD2", "CE1": "CE2"},
-    "TYR": {"CD1": "CD2", "CE1": "CE2"},
-}
-
-
-@cache
-def _make_ambiguous_atom_swap_indices() -> list[list]:
-    out = []
-    for res in STANDARD_RESIDUES:
-        atoms = ATOM_THIN_ATOMS[AMINO_ACID_1_TO_3[res]]
-        res_swaps = _AMBIGUOUS_ATOM_SWAPS.get(AMINO_ACID_1_TO_3[res], {})
-        res_swaps = res_swaps | dict((v, k) for k, v in res_swaps.items())
-        indices = [
-            atoms.index(res_swaps.get(a, a)) if a else i for i, a in enumerate(atoms)
-        ]
-        out.append(indices)
-    return out
 
 
 def _gapped_domain_to_numbering(
@@ -542,29 +518,6 @@ class ResidueArray:
             for k, v in gapped.items()
         }
         return self.renumber(numbering)
-
-    def swap_ambiguous_atoms(self):
-        ambiguous_swap_indices = torch.as_tensor(
-            _make_ambiguous_atom_swap_indices(), device=self.residue_type.device
-        )
-
-        ambiguous_swap_indices = ambiguous_swap_indices[self.residue_type]
-
-        atom_thin_xyz = torch.gather(
-            self.atom_thin_xyz,
-            dim=-2,
-            index=einops.repeat(ambiguous_swap_indices, "... -> ... 3"),
-        )
-
-        atom_thin_mask = torch.gather(
-            self.atom_thin_mask,
-            dim=-1,
-            index=ambiguous_swap_indices,
-        )
-
-        return dataclasses.replace(
-            self, atom_thin_xyz=atom_thin_xyz, atom_thin_mask=atom_thin_mask
-        )
 
     def kabsch(
         self,
